Renamed worker.common to worker.tasks
