+++ src/pytest_celery/api/setup.py
@@ -20,7 +20,7 @@ class CeleryTestSetup:
         self._backend_cluster = backend_cluster
         self._app = app
 
-        from pytest_celery.components.worker.common import ping
+        from pytest_celery.components.worker.tasks import ping
 
         self.ping = ping
